Update books/05_Deuteronomy.tex
Numerous changes in word order, tense, etc. as noted:

Deut 16:6 – There is an extra/unnecessary comma at the end of the verse.
 
Deut 16:14 – “…you, your son, your daughter, your male and female slaves, the descendants of Levi, foreigners, orphans, and widows, who live in your cities.” – I believe there should be an “and” before “the descendants of Levi”.  The section from “the descendants of Levi” to “widows” is a secondary nested serial list.
Changed as recommended in Jutoh ebook matrix, Microsoft Word DOC file, and Braille text file.
Deut 18:1 – Should there be en-dashes before and after “the whole tribe of Levi” rather than commas?
Yes, but they're em-dashes, not en-dashes (Code ALT-0151)
Deut 18:4 – “Give them the first gatherings of your grain, wine, oil, and wool from the shearing of your flock.” – I’m not entirely sure what to suggest, but this is currently somewhat awkward as rendered.  It reads as if the grain, wine, oil, and wool all come from the shearing of the flock.  All other versions distinguish the grain, wine, and oil from the wool by some means or another.
Changed to "...wine, and oil, as well as wool..."
Deut 18:22 – I believe the comma after “LORD” is unnecessary.
Deleted
Deut 19:6 – This verse is clumsy and doesn’t make sense as written…especially the first section.  I’m not entirely sure what to suggest, but if it doesn’t violate the text or change the meaning, then something along the lines of this would make much more sense: “Since the distance may be great, the angry avenger may overtake the killer he is pursuing and kill him, in which case there will be no justice in his death, because he did not hate his friend previously.”
Changed as suggested.
Deut 21:5 – There is an extra space between “of” and “the” near the beginning of the verse.
Changed as suggested.
Deut 21:8 – Shouldn’t “bloodshed” be one word?  (cf. 17:8 and 22:8)
Changed to "blood that has been shed"
 
Deut 22:1-4 – There seem to be number agreement problems through this whole paragraph.  – “When you see the ox (single) or sheep (single)…don’t…leave them (plural)…be sure to return them (plural)…bring the animal (single)…let it (single) remain…return it (single)…etc” – I believe that the two instances of “them” in v.1 and the two instances of “them” in v.4 should all be changed to “it” in agreement with the rest of the paragraph.
"Them" changed to "it" as suggested.
Deut 22:7 – I believe there should be a comma after “young”
 
Deut 22:12 – I’m not at all sure of this one, but should “yourselves” be changed to “yourself” in agreement with the end of the sentence??
Yourselves changed to yourself.
Deut 22:13 – Should there be a comma after “wife”?
 
Deut 22:19 – “…because he had defamed a virgin…” – verb tense agreement – I believe it should be “…because he has defamed a virgin…” in agreement with the verb tense of the rest of the passage.
Changed to "he defamed"
Deut 22:20 – “…virginity weren’t found…” – verb tense agreement – I believe it should be “…virginity isn’t found…” (or perhaps “wasn’t”?)
"weren't" changed to "wasn't"
Deut 22:23 – “If a man meets a young virgin lady engaged to be married in the city…” – I believe the sentence order should be “If a man meets a young virgin lady in the city who is engaged to be married…” (cf. the similar construction in v.25)
Changed to "lady in the city who is engaged to be married..."
Deut 22:29 – “Because he had violated her…” – verb tense agreement – should be “Because he has violated her…”
Changed to "he violated".

Amended by Charles Welty. Authorized by William P. Welty, Ph.D., of The ISV Foundation on 06 JAN 2013 at 9:40AM PST.